ResponseEntity, Spring Validation

+++ SB-02-RestApis/src/main/java/com/tyss/service/ProductServiceImp.java
@@ -10,6 +10,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.tyss.dao.ProductDAO;
@@ -27,23 +29,32 @@ public class ProductServiceImp implements ProductService {
 	private ProductDAO productDAO;
 
 	@Override
-	public String save(ProductDTO productDTO) {
+	public ResponseEntity<String> save(ProductDTO productDTO) {
 		Product product = new Product();
 		BeanUtils.copyProperties(productDTO, product);
 		Product savedProduct = productRepository.save(product);
-		return "Product is saved with Id : " + savedProduct.getPid();
+
+//		ResponseEntity<String> re = new ResponseEntity<String>(
+//				"Product is saved with Id : " + savedProduct.getPid(),
+//				HttpStatus.CREATED);
+
+		return ResponseEntity.status(HttpStatus.CREATED).body("Product is saved with Id : " + savedProduct.getPid());
 	}
 
 	@Override
-	public Product findById(Integer pid) {
+	public ResponseEntity<?> findById(Integer pid) {
 		Optional<Product> opt = productRepository.findById(pid);
 
 //		if (opt.isPresent()) {
 //			return opt.get();
 //		}
 
-		Product product = opt.orElseThrow(() -> new RuntimeException("Product not found"));
-		return product;
+//		Product product = opt.orElseThrow(() -> new RuntimeException("Product not found"));
+		if (opt.isPresent()) {
+			return new ResponseEntity<Product>(opt.get(), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<String>("Product not found", HttpStatus.NOT_FOUND);
+		}
 	}
 
 	// pagination
@@ -80,9 +91,13 @@ public class ProductServiceImp implements ProductService {
 	}
 
 	@Override
-	public String deleteById(Integer pid) {
-		productRepository.deleteById(pid);
-		return "Deleted Successfully";
+	public ResponseEntity<String> deleteById(Integer pid) {
+		if (productRepository.findById(pid).isPresent()) {
+			productRepository.deleteById(pid);
+			return ResponseEntity.ok("Deleted Successfully");
+		} else {
+			return new ResponseEntity<String>("Not found", HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@Override
@@ -93,6 +108,12 @@ public class ProductServiceImp implements ProductService {
 	@Override
 	public String updatePrice(Integer pid, Double price) {
 		return productDAO.updatePrice(pid, price);
+	}
+
+	@Override
+	public ResponseEntity<List<Product>> getByNameContaining(String name) {
+		List<Product> products = productRepository.findByNameContainingAllIgnoringCase(name);
+		return ResponseEntity.ok().body(products);
 	}
 
 }
